Use `case` for unprotect_reviews
The case considers all known valid boolean YAML values for True and
False.
If an invalid value is given, the choice falls back to the default
(False).

The change in action.yml is only aesthetic, making the choices' keys in
similar in order for all choices.

Add requirements*.txt to Dockerfile.