New release workflow
The new workflow will run on publishing a new release version.

It will run `invoke update-version` passing the tag name of the new
release, which will update `push_action/__init__.py:__version__`.
Then, it will proceed to commit this change, update the newly created
tag, as well as update the `vMAJOR` tag to this newly created commit.

Finally, it will use the local action to push to `master`.