Add fail_fast CI job
To ensure fail_fast actually works as intended, one needs to manually
check the output in the CI job to see if it truly works and doesn't
ignore the option.

In order to automate this, the best way I see is to store everything
reported to stderr as an action output - but that seems like overkill
just to be able to do testing?